requirements.txt
spacy==3.7.4
textstat==0.7.4
requests==2.32.2
pandas==2.0.3
openai==1.37.1
openpyxl==3.1.2
python-dotenv==1.0.1
matplotlib==3.7.3
seaborn==0.13.2